Refactor: Update ProductContext and Cart components
- Update ProductContext to use localStorage for saved items and cart
- Remove unused Order component and replace it with OrderForm in Cart component
- Add OrderForm component to handle placing orders
- Implement placeOrder function to send order details to the server
- Display success and error toasts when placing an order
