--- src/pages/Cart/Order.jsx
@@ -1,5 +1,0 @@
-export const Order = () => {
-  return (
-    <div>Order</div>
-  )
-}
